Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/block-all-mixed-content/insecure-image-in-iframe-with-enforced-and-report-policies.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
    testRunner.dumpFrameLoadCallbacks();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<p>This test loads a secure iframe that loads an insecure image.  We should trigger a
mixed content block because the child frame has an CSP directive block-all-mixed-content
in an enforced policy.</p>
<iframe src="https://127.0.0.1:8443/security/contentSecurityPolicy/block-all-mixed-content/resources/frame-with-insecure-image-with-enforced-and-report-policies.php" width="100%" height="300"></iframe>
</body>
</html>
